extract create-lane functionality to a new file
